Draft for refactored network settings
