+++ sdk/ml/azure-ai-ml/azure/ai/ml/_schema/_utils/utils.py
@@ -23,19 +23,6 @@ def validate_arm_str(arm_str: str) -> bool:
     raise ValidationError(f"ARM string {arm_str} is not formatted correctly.")
 
 
-def get_subnet_str(vnet_name: str, subnet: str, sub_id: Optional[str] = None, rg: Optional[str] = None) -> str:
-    if vnet_name and not subnet:
-        raise ValidationError("Subnet is required when vnet name is specified.")
-    try:
-        if validate_arm_str(subnet):
-            return subnet
-    except ValidationError:
-        return (
-            f"/subscriptions/{sub_id}/resourceGroups/{rg}/"
-            f"providers/Microsoft.Network/virtualNetworks/{vnet_name}/subnets/{subnet}"
-        )
-
-
 def replace_key_in_odict(odict: OrderedDict, old_key: Any, new_key: Any):
     if not odict or old_key not in odict:
         return odict
