+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_compute/_network_settings.py
@@ -1,0 +1,69 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+
+# pylint: disable=protected-access
+
+from typing import Optional
+from marshmallow.exceptions import ValidationError
+
+from azure.ai.ml._schema._utils.utils import validate_arm_str
+from azure.ai.ml.exceptions import ErrorCategory, ErrorTarget, ValidationException
+
+
+class NetworkSettings:
+    def __init__(
+        self,
+        *,
+        subnet: Optional[str] = None,
+        vnet_name: Optional[str] = None,
+        **kwargs,
+    ):
+        """Network settings for a compute.
+
+        :param subnet: The subnet name. Can also be provided as an ARM string specifying the location of the subnet. Defaults to None
+        :type subnet: str, optional
+        :param vnet_name: The virtual network name. Ignored if subnet is provided as an ARM string. Defaults to None.
+        :type vnet_name: str, optional
+
+        """
+        if vnet_name and not subnet:
+            msg = "Subnet is required when vnet name is specified."
+            raise ValidationException(
+                message=msg,
+                target=ErrorTarget.COMPUTE,
+                no_personal_data_message=msg,
+                error_category=ErrorCategory.USER_ERROR,
+            )
+        self.subnet = subnet
+        self.vnet_name = vnet_name
+        self._public_ip_address = kwargs.pop("public_ip_address", None)
+        self._private_ip_address = kwargs.pop("private_ip_address", None)
+
+    @property
+    def public_ip_address(self) -> str:
+        """Public IP address of the compute instance.
+
+        return: Public IP address.
+        rtype: str
+        """
+        return self._public_ip_address
+
+    @property
+    def private_ip_address(self) -> str:
+        """Private IP address of the compute instance.
+
+        return: Private IP address.
+        rtype: str
+        """
+        return self._private_ip_address
+
+    def get_subnet_str(self, sub_id: Optional[str] = None, rg: Optional[str] = None) -> str:
+        try:
+            if validate_arm_str(self.subnet):
+                return self.subnet
+        except ValidationError:
+            return (
+                f"/subscriptions/{sub_id}/resourceGroups/{rg}/"
+                f"providers/Microsoft.Network/virtualNetworks/{self.vnet_name}/subnets/{self.subnet}"
+            )
