+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_compute/compute.py
@@ -197,42 +197,3 @@ class Compute(Resource, RestTranslatableMixin):
     @abstractclassmethod
     def _load_from_dict(cls, data: Dict, context: Dict, **kwargs) -> "Compute":
         pass
-
-
-class NetworkSettings:
-    def __init__(
-        self,
-        *,
-        vnet_name: Optional[str] = None,
-        subnet: Optional[str] = None,
-        **kwargs,
-    ):
-        """Network settings for a compute.
-
-        :param vnet_name: The virtual network name, defaults to None
-        :type vnet_name: str, optional
-        :param subnet: The subnet name, defaults to None
-        :type subnet: str, optional
-        """
-        self.vnet_name = vnet_name
-        self.subnet = subnet
-        self._public_ip_address = kwargs.pop("public_ip_address", None)
-        self._private_ip_address = kwargs.pop("private_ip_address", None)
-
-    @property
-    def public_ip_address(self) -> str:
-        """Public IP address of the compute instance.
-
-        return: Public IP address.
-        rtype: str
-        """
-        return self._public_ip_address
-
-    @property
-    def private_ip_address(self) -> str:
-        """Private IP address of the compute instance.
-
-        return: Private IP address.
-        rtype: str
-        """
-        return self._private_ip_address
