+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_compute/compute_instance.py
@@ -20,11 +20,10 @@ from azure.ai.ml._restclient.v2022_10_01_preview.models import (
     PersonalComputeInstanceSettings,
     ResourceId,
 )
-from azure.ai.ml._schema._utils.utils import get_subnet_str
 from azure.ai.ml._schema.compute.compute_instance import ComputeInstanceSchema
 from azure.ai.ml.constants._common import BASE_PATH_CONTEXT_KEY, TYPE
 from azure.ai.ml.constants._compute import ComputeDefaults, ComputeType
-from azure.ai.ml.entities._compute.compute import Compute, NetworkSettings
+from azure.ai.ml.entities._compute.compute import Compute
 from azure.ai.ml.entities._credentials import IdentityConfiguration
 from azure.ai.ml.entities._mixins import DictMixin
 from azure.ai.ml.entities._util import load_from_dict
@@ -33,6 +32,7 @@ from ._image_metadata import ImageMetadata
 from ._schedule import ComputeSchedules
 from ._setup_scripts import SetupScripts
 from ._custom_applications import CustomApplications, validate_custom_applications
+from ._network_settings import NetworkSettings
 
 module_logger = logging.getLogger(__name__)
 
@@ -299,9 +299,7 @@ class ComputeInstance(Compute):
 
     def _set_full_subnet_name(self, subscription_id: str, rg: str) -> None:
         if self.network_settings:
-            self.subnet = get_subnet_str(
-                self.network_settings.vnet_name,
-                self.network_settings.subnet,
+            self.subnet = self.network_settings.get_subnet_str(
                 subscription_id,
                 rg,
             )
